dbt/scripts/requirements.txt
asn1crypto==1.5.1
boto3==1.38.41
botocore==1.38.41
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
filelock==3.18.0
idna==3.10
jmespath==1.0.1
packaging==25.0
platformdirs==4.3.8
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
s3transfer==0.13.0
six==1.17.0
snowflake-connector-python==3.15.0
sortedcontainers==2.4.0
tomlkit==0.13.3
typing_extensions==4.14.0
urllib3==2.5.0
